+++ src/app/vendors/SingleVendor.tsx
@@ -10,7 +10,7 @@ import { useEventBusListener } from 'app/utils/hooks';
 import { DestinyCollectibleComponent, DestinyProfileResponse } from 'bungie-api-ts/destiny2';
 import clsx from 'clsx';
 import React, { useCallback, useEffect } from 'react';
-import { connect } from 'react-redux';
+import { connect, useSelector } from 'react-redux';
 import { useLocation, useParams } from 'react-router';
 import { DestinyAccount } from '../accounts/destiny-account';
 import Countdown from '../dim-ui/Countdown';
@@ -28,7 +28,11 @@ import { refresh$ } from '../shell/refresh-events';
 import { loadAllVendors } from './actions';
 import { toVendor } from './d2-vendors';
 import type { VendorsState } from './reducer';
-import { mergedCollectiblesSelector, vendorsByCharacterSelector } from './selectors';
+import {
+  mergedCollectiblesSelector,
+  ownedVendorItemsSelector,
+  vendorsByCharacterSelector,
+} from './selectors';
 import styles from './SingleVendor.m.scss';
 import { VendorLocation } from './Vendor';
 import VendorItems from './VendorItems';
@@ -40,7 +44,6 @@ interface ProvidedProps {
 interface StoreProps {
   stores: DimStore[];
   buckets?: InventoryBuckets;
-  ownedItemHashes: Set<number>;
   profileResponse?: DestinyProfileResponse;
   vendors: VendorsState['vendorsByCharacter'];
   defs?: D2ManifestDefinitions;
@@ -70,7 +73,6 @@ function SingleVendor({
   account,
   stores,
   buckets,
-  ownedItemHashes,
   profileResponse,
   dispatch,
   vendors,
@@ -89,6 +91,10 @@ function SingleVendor({
   if (!characterId) {
     throw new Error('no characters chosen or found to use for vendor API call');
   }
+
+  const ownedItemHashes = useSelector((state: RootState) =>
+    ownedVendorItemsSelector(state, characterId)
+  );
 
   const vendorData = characterId ? vendors[characterId] : undefined;
   const vendorResponse = vendorData?.vendorsResponse;